+++ src/net/fe/network/message/QuitMessage.java
@@ -4,7 +4,7 @@ import net.fe.network.Message;
 
 public class QuitMessage extends Message {
 	private static final long serialVersionUID = 6558427589809693714L;
-	public QuitMessage(int origin) {
+	public QuitMessage(byte origin) {
 		super(origin);
 	}
